Fix gzip use bug
For some reason, gzip needs the file reading to it as a file object so
you have to first open the file and then read this into the gzip file to
compress. The way I had it before took the string of the file location
e.g. /tmp/2014-D-42.gpx and compressed this string, not the file it
refers to.

Now try to get it to work with the with open file as notation